+++ packages/composable-controller/src/ComposableController.ts
@@ -126,8 +126,10 @@ export class ComposableController<
         StateMetadata<ComposableControllerState>
       >((metadata, name) => {
         (metadata as StateMetadataConstraint)[name] = {
+          includeInStateLogs: false,
           persist: true,
           anonymous: true,
+          usedInUi: false,
         };
         return metadata;
       }, {} as never),